backend/Farmers_market_backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.2
